refactor: re-structure input and field components
