--- src/components/TextField/styled.ts
@@ -1,12 +1,0 @@
-import styled from 'styled-components';
-
-export const FieldWrapper = styled.div`
-  margin: 0 2.5vw;
-  width: 300px;
-  margin: 10px;
-`;
-
-export const Input = styled.input`
-  display: block;
-  width: 100%;
-`;
