+++ src/components/fields/FieldWrapper/index.ts
@@ -1,0 +1,3 @@
+import { FieldWrapper } from './styled';
+
+export default FieldWrapper;
