+++ src/components/inputs/Input/index.ts
@@ -1,0 +1,3 @@
+import { Input } from './styled';
+
+export default Input;
