+++ src/components/inputs/Input/styled.ts
@@ -1,11 +1,5 @@
 import styled from 'styled-components';
 
-export const FieldWrapper = styled.div`
-  margin: 0 2.5vw;
-  width: 300px;
-  margin: 10px;
-`;
-
 export const Input = styled.input`
   display: block;
   width: 100%;
