+++ src/components/inputs/Select/index.ts
@@ -1,0 +1,3 @@
+import { Select } from './styled';
+
+export default Select;
