+++ src/components/inputs/Select/styled.ts
@@ -1,0 +1,6 @@
+import styled from 'styled-components';
+
+export const Select = styled.select`
+  display: block;
+  width: 100%;
+`;
