+++ src/components/inputs/TextArea/index.ts
@@ -1,0 +1,3 @@
+import { TextArea } from './styled';
+
+export default TextArea;
