+++ src/components/inputs/TextArea/styled.ts
@@ -1,11 +1,5 @@
 import styled from 'styled-components';
 
-export const FieldWrapper = styled.div`
-  margin: 0 2.5vw;
-  width: 300px;
-  margin: 10px;
-`;
-
 export const TextArea = styled.textarea`
   display: block;
   width: 100%;
